LPS-83992 Merge EventAnnotationImpl, EventDescriptorImpl and BaseEventImpl into EventImpl
